langgraphs/langgraph_ex02.py
from langgraph.graph import StateGraph, END, START
from typing_extensions import TypedDict, Annotated
from langgraph.graph.message import add_messages


class State(TypedDict):
    messages: Annotated[list, add_messages]


workflow = StateGraph(State)




from IPython.display import Image, display

try:
    display(Image(workflow.get_graph().draw_mermaid_png()))
except Exception:
    # This requires some extra dependencies and is optional
    pass