pyproject.toml
[project]
name = "study"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi[all]>=0.115.14",
    "ipython>=9.4.0",
    "langchain-community>=0.3.27",
    "langchain-openai>=0.3.27",
    "langgraph>=0.5.1",
    "python-json-logger>=3.3.0",
    "requests>=2.32.4",
    "wikipedia>=1.4.0",
    "wikipedia-api>=0.8.1",
]
